refactor(frontend): refactor `ChannelList` viewModel
