+++ frontend/src/pages/main/channel/_components/channel-list/channel-list.stories.tsx
@@ -1,11 +1,9 @@
+import { createSignal } from 'solid-js';
 import { StoryFn } from 'storybook-solidjs';
 
 import * as styles from './channel-list.stories.css';
-import { ChannelList, ChannelListProps, ChannelListViewModelType } from './channel-list';
+import { ChannelList, ChannelListProps } from './channel-list';
 
-import IconSearch from '@/assets/icons/search.svg';
-import IconAddChat from '@/pages/main/channel/_assets/icons/add-chat.svg';
-import { createSignal } from 'solid-js';
 import { ListProfile } from '../../_types/channel-list-item';
 
 export default {
@@ -20,47 +18,36 @@ const Template: StoryFn<ChannelListProps> = () => {
       profileUrl: `https://picsum.photos/200?s=${Math.random()}`,
     }] as [string, ListProfile]);
 
-  const NoOpViewModel: ChannelListViewModelType = () => ({
-    channels: () => Array.from({ length: 10 }).map((_, index) => {
-      const userLength = 2 + Math.floor(Math.random() * 10);
-      const displayUsers = getUsers(userLength);
+  const channelList = () => Array.from({ length: 10 }).map((_, index) => {
+    const userLength = 2 + Math.floor(Math.random() * 10);
+    const displayUsers = getUsers(userLength);
 
-      const lastChat = {
-        chatType: 0,
-        nickname: displayUsers[Math.floor(Math.random() * userLength)][1].nickname,
-        content: 'last message',
-        timestamp: new Date(),
-      };
+    const lastChat = {
+      chatType: 0,
+      nickname: displayUsers[Math.floor(Math.random() * userLength)][1].nickname,
+      content: 'last message',
+      timestamp: new Date(),
+    };
 
-      return {
-        id: `${index}`,
-        displayUsers,
+    return {
+      id: `${index}`,
+      displayUsers,
 
-        lastChat: Math.random() > 0.2 ? lastChat : undefined,
+      lastChat: Math.random() > 0.2 ? lastChat : undefined,
 
-        name: `Channel ${index + 1}`,
-        unreadCount: Math.floor(Math.random() * 10),
-        userCount: displayUsers.length,
-        silent: Math.random() > 0.5,
-      };
-    }),
-    topItems: () => [
-      {
-        kind: 'click',
-        icon: <IconSearch />,
-      },
-      {
-        kind: 'click',
-        icon: <IconAddChat />,
-      },
-    ],
+      name: `Channel ${index + 1}`,
+      unreadCount: Math.floor(Math.random() * 10),
+      userCount: displayUsers.length,
+      silent: Math.random() > 0.5,
+    };
   });
+
   const [activeId, setActiveId] = createSignal('1');
 
   return (
     <div class={styles.background}>
       <ChannelList
-        viewModel={NoOpViewModel}
+        channels={channelList()}
         activeId={activeId()}
         setActiveId={setActiveId}
       />
