+++ frontend/src/pages/main/channel/_hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useChannelList';
